chore(table): seperate types from the code
