+++ package.json
@@ -31,6 +31,7 @@
     "./hooks/*": "./src/hooks/*.tsx",
     "./atoms/*": "./@/components/ui/*.tsx",
     "./molecules/*": "./src/molecules/*/index.tsx",
+    "./molecules/*/types": "./src/molecules/*/types.ts",
     "./organisms/*": "./src/organisms/*/index.tsx",
     "./organisms/**/*": "./src/organisms/**/*/index.tsx",
     "./organisms/*/types": "./src/organisms/*/types/index.ts",
